Move widely used types into types file
