+++ src/components/about/index.js
@@ -1,5 +1,3 @@
-import {h} from "preact";
-
 const About = () => (
     <div class="page">
         <h2 class="leftColumn">About ICAT Admin</h2>
